refactor connection/disconnection notifications
Instead of the server sending a 'status' notification with a message,
it now sends 'connected' and 'disconnected' messages, and leaves
choosing a message to the client.

Note that for this to work, I had to patch the socket.io-client library:

```
--- socket.io.js.orig   2020-10-22 07:53:02.258582031 +0200
+++ socket.io.js        2020-10-22 07:48:45.877731503 +0200
@@ -3660,7 +3660,7 @@
   if (global.attachEvent) {
     global.attachEvent('onunload', unloadHandler);
   } else if (global.addEventListener) {
-    global.addEventListener('beforeunload', unloadHandler, false);
+    global.addEventListener('unload', unloadHandler, false); //was: beforeunload
   }
 }
```

Without this change, the `beforeunload` handler there would run and remove
various `Request` objects that are still needed in case the window isn't
closed.

This seems to be fixed in the latest version, so perhaps we should update:

https://github.com/socketio/socket.io/blob/129c6417bd818bc8b4e1b831644323876e627c13/client-dist/socket.io.js#L3925-L3926